// src/store/interfaces.ts
export type ConversionUnit = {
  label: string;
  multiplier: number;
  offset: number;
  type: string;
};

export type ConversionUnits = {
  [key: string]: ConversionUnit;
};